Fixes to allow AIX to build and run, and some other bugfixes
Type exception initialization was broken, accessing the wrong exception
variable.
AIX needs the alloca header pulled in as well, so this was changed.
AIX has some undefined things that shouldn't stop building, so the
makefile was changed to not set -Werror or -Wundef on AIX.
GCC on AIX64 doesn't set the ppc64 defines, so an extra check in the ppc
section is necessary to detect 64-bit.