Raise a better error for read-only or noexec temp.
When the jffi stub library must be emitted to and loaded from a
local filesystem, we may end up doing so with a path that is
either read-only or "noexec". This change attempts to detect those
cases and inform users about why the error happened and how to
fix it.

See jruby/jruby#5900.