Don't attempt to initCause this exception.
In jruby/jruby#1302 we see that attempting to load FFI type defs
when the jffi stub can't be loaded results in
IllegalStateException being raised when we try to initCause on an
UnsatisfiedLinkError. It appears that (sometimes?) simply
constructing this exception initializes its cause, which can only
be done once. Our attempt to initialize it again triggers the
error.

Since ULE does not have a constructor that accepts a cause, the
only option here is to include the message from the triggering
exception.