Add a property to tell jffi where to extract lib
This adds the ability to specify the target directory for the
native library rather than using the standard tmp location with
fallback to the current directory. If the property
jffi.extract.dir is provided, that directory and only that
directory will be used when attempting to extract and load the
jarred native library for the current platform. The file lifecycle
still follows that of java.lang.File.createTempFile.