Add another constructor for JRuby FFI use
JRuby FFI duplicates some logic, notably logic from
VariadicInvoker, due to private or package-private utilities
in jnr-ffi. In this case, it needs to be able to set up the
CallContext for the function using the fixed param count, but
there's no constructor to do this and no utility function to
create the CallContext manually.

This duplication will need to be cleaned up but that is out of
scope for this change. I add this constructor so that the duped
code in JRuby can make the proper call.